joker_backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.2.11
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
executing==2.0.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.43
psycopg[binary]
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.2
typing-extensions==4.10.0
wcwidth==0.2.13
